pyproject.toml
[project]
name = "pr2m3"
version = "0.1.0"
description = ""
authors = [
    {name = "'Ruslan",email = "'sounditbox@gmail.com'"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ipython (>=9.3.0,<10.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
